+++ src/fromString/fromString.js
@@ -4,16 +4,18 @@ function fromString(parser, defaultLocale) {
 		if (parsed.invalid) {
 			return parsed;
 		}
-		const date = new Date();
-		// default to current year, month and day
+		// set to Jan 1 to prevent setting non-existent days
+		const date = new Date(2000, 0, 1);
+		// first setting day, then year, then month
+		// to handle months with fewer days and years without Feb 29
+		if (typeof parsed.day === 'number') {
+			date.setUTCDate(parsed.day);
+		}
 		if (typeof parsed.year === 'number') {
 			date.setUTCFullYear(parsed.year);
 		}
 		if (typeof parsed.month === 'number') {
 			date.setUTCMonth(parsed.month - 1);
-		}
-		if (typeof parsed.day === 'number') {
-			date.setUTCDate(parsed.day);
 		}
 		// default to first unit for time components
 		date.setUTCHours(parsed.hour || 0);