+++ src/fromString/fromString.spec.js
@@ -1,6 +1,5 @@
 const fromString = require('./fromString.js');
 
-const now = new Date();
 describe('fromString', () => {
 	it('should return invalid dates', () => {
 		const invalid = { invalid: 'foo' };
@@ -8,20 +7,18 @@ describe('fromString', () => {
 		const fromFn = fromString(parser);
 		expect(fromFn()).toBe(invalid);
 	});
-	it('should use current month and day', () => {
+	it('should default to Jan 1', () => {
 		const result = { year: 2020 };
 		const parser = { attempt: jest.fn(() => result) };
 		const fromFn = fromString(parser);
-		const jan1 = new Date(
-			Date.UTC(2020, now.getUTCMonth(), now.getUTCDate(), 0, 0, 0, 0)
-		);
+		const jan1 = new Date(Date.UTC(2020, 0, 1, 0, 0, 0, 0));
 		expect(fromFn()).toEqual(jan1);
 	});
-	it('should use current month', () => {
+	it('should default to 1st of month', () => {
 		const result = { year: 2020, month: 5 };
 		const parser = { attempt: jest.fn(() => result) };
 		const fromFn = fromString(parser);
-		const may1 = new Date(Date.UTC(2020, 4, now.getUTCDate(), 0, 0, 0, 0));
+		const may1 = new Date(Date.UTC(2020, 4, 1, 0, 0, 0, 0));
 		expect(fromFn()).toEqual(may1);
 	});
 	it('should reset all but day', () => {